Generate constants from GANA recfiles.

+++ src/gnunet/message/msg_core.go
@@ -60,7 +60,7 @@ func NewEphemeralKeyMsg() *EphemeralKeyMsg {
 		SignedBlock: &EphKeyBlock{
 			Purpose: &crypto.SignaturePurpose{
 				Size:    88,
-				Purpose: enums.SIG_SET_ECC_KEY,
+				Purpose: uint32(enums.SIG_SET_ECC_KEY),
 			},
 			CreateTime:   util.AbsoluteTimeNow(),
 			ExpireTime:   util.NewRelativeTime(12 * time.Hour),
