+++ src/gnunet/message/msg_transport.go
@@ -145,7 +145,7 @@ func NewSignedAddress(a *util.Address) *SignedAddress {
 	addr := &SignedAddress{
 		Purpose: &crypto.SignaturePurpose{
 			Size:    uint32(alen + 20),
-			Purpose: enums.SIG_TRANSPORT_PONG_OWN,
+			Purpose: uint32(enums.SIG_TRANSPORT_PONG_OWN),
 		},
 		ExpireOn: util.AbsoluteTimeNow().Add(12 * time.Hour),
 		AddrSize: uint32(alen),
